# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "GoTunnel 内网穿透服务器 API",
        "title": "GoTunnel API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:7500",
    "basePath": "/",
    "paths": {
...
        "/api/client/{id}": {
            "get": {
                "description": "获取指定客户端的详细信息",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "客户端"
                ],
                "summary": "获取客户端详情",
                "parameters": [
                    {
                        "type": "string",
                        "description": "客户端ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/handler.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.ClientResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            },
            "put": {
                "description": "更新指定客户端的配置信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "客户端"
                ],
                "summary": "更新客户端配置",
                "parameters": [
                    {
                        "type": "string",
                        "description": "客户端ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "更新内容",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateClientRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            },
            "delete": {
                "description": "删除指定的客户端配置",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "客户端"
                ],
                "summary": "删除客户端",
                "parameters": [
                    {
                        "type": "string",
                        "description": "客户端ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/client/{id}/disconnect": {
            "post": {
                "description": "强制断开客户端连接",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "客户端"
                ],
                "summary": "断开连接",
                "parameters": [
                    {
                        "type": "string",
                        "description": "客户端ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
...
        "/api/client/{id}/push": {
            "post": {
                "description": "将配置推送到在线客户端",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "客户端"
                ],
                "summary": "推送配置",
                "parameters": [
                    {
                        "type": "string",
                        "description": "客户端ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/client/{id}/restart": {
            "post": {
                "description": "发送重启命令到客户端",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "客户端"
                ],
                "summary": "重启客户端",
                "parameters": [
                    {
                        "type": "string",
                        "description": "客户端ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/client/{id}/screenshot": {
            "get": {
                "description": "获取在线客户端当前屏幕截图",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "客户端"
                ],
                "summary": "获取客户端截图",
                "parameters": [
                    {
                        "type": "string",
                        "description": "客户端ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "JPEG 质量，1-100，0 使用默认值",
                        "name": "quality",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/handler.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.ScreenshotResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/client/{id}/shell": {
            "post": {
                "description": "在在线客户端执行单条 Shell 命令并返回输出",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "客户端"
                ],
                "summary": "执行客户端 Shell 命令",
                "parameters": [
                    {
                        "type": "string",
                        "description": "客户端ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Shell 执行参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.ExecuteShellRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/handler.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.ShellExecuteResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/client/{id}/system-stats": {
            "get": {
                "description": "获取在线客户端的系统资源使用情况",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "客户端"
                ],
                "summary": "获取客户端系统状态",
                "parameters": [
                    {
                        "type": "string",
                        "description": "客户端ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/handler.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.SystemStatsResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/clients": {
            "get": {
                "description": "返回所有注册客户端的列表及其在线状态",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "客户端"
                ],
                "summary": "获取所有客户端",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/handler.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/dto.ClientListItem"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            },
            "post": {
                "description": "创建一个新的客户端配置",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "客户端"
                ],
                "summary": "创建新客户端",
                "parameters": [
                    {
                        "description": "客户端信息",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateClientRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
...
    },
    "definitions": {
...
        "dto.ClientListItem": {
            "description": "客户端列表中的单个项目",
            "type": "object",
            "properties": {
                "arch": {
                    "type": "string",
                    "example": "amd64"
                },
                "id": {
                    "type": "string",
                    "example": "client-001"
                },
                "last_offline_at": {
                    "type": "integer"
                },
                "last_ping": {
                    "type": "string"
                },
                "nickname": {
                    "type": "string",
                    "example": "My Client"
                },
                "online": {
                    "type": "boolean",
                    "example": true
                },
                "os": {
                    "type": "string",
                    "example": "linux"
                },
                "remote_addr": {
                    "type": "string"
                },
                "rule_count": {
                    "type": "integer",
                    "example": 3
                },
                "version": {
                    "type": "string",
                    "example": "1.0.0"
                }
            }
        },
        "dto.ClientResponse": {
            "description": "客户端详细信息",
            "type": "object",
            "properties": {
                "arch": {
                    "type": "string",
                    "example": "amd64"
                },
                "id": {
                    "type": "string",
                    "example": "client-001"
                },
                "last_offline_at": {
                    "type": "integer",
                    "example": 1735785000
                },
                "last_ping": {
                    "type": "string",
                    "example": "2025-01-02T10:30:00Z"
                },
                "nickname": {
                    "type": "string",
                    "example": "My Client"
                },
                "online": {
                    "type": "boolean",
                    "example": true
                },
                "os": {
                    "type": "string",
                    "example": "linux"
                },
                "remote_addr": {
                    "type": "string",
                    "example": "192.168.1.100:54321"
                },
                "rules": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.ProxyRule"
                    }
                },
                "version": {
                    "type": "string",
                    "example": "1.0.0"
                }
            }
        },
...
        "dto.CreateClientRequest": {
            "description": "创建新客户端的请求体",
            "type": "object",
            "required": [
                "id"
            ],
            "properties": {
                "id": {
                    "type": "string",
                    "maxLength": 64,
                    "minLength": 1,
                    "example": "client-001"
                },
                "rules": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.ProxyRule"
                    }
                }
            }
        },
        "dto.ExecuteShellRequest": {
            "type": "object",
            "required": [
                "command"
            ],
            "properties": {
                "command": {
                    "type": "string"
                },
                "timeout": {
                    "type": "integer"
                }
            }
        },
...
        "dto.ProxyRule": {
            "type": "object",
            "properties": {
                "auth_enabled": {
                    "type": "boolean"
                },
                "auth_password": {
                    "type": "string"
                },
                "auth_username": {
                    "type": "string"
                },
                "enabled": {
                    "type": "boolean"
                },
                "local_ip": {
                    "type": "string"
                },
                "local_port": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "port_status": {
                    "type": "string"
                },
                "remote_port": {
                    "type": "integer"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "dto.ScreenshotResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "string"
                },
                "error": {
                    "type": "string"
                },
                "height": {
                    "type": "integer"
                },
                "timestamp": {
                    "type": "integer"
                },
                "width": {
                    "type": "integer"
                }
            }
        },
...
        "dto.ShellExecuteResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "exit_code": {
                    "type": "integer"
                },
                "output": {
                    "type": "string"
                }
            }
        },
...
        "dto.SystemStatsResponse": {
            "type": "object",
            "properties": {
                "cpu_usage": {
                    "type": "number"
                },
                "disk_total": {
                    "type": "integer"
                },
                "disk_usage": {
                    "type": "number"
                },
                "disk_used": {
                    "type": "integer"
                },
                "memory_total": {
                    "type": "integer"
                },
                "memory_usage": {
                    "type": "number"
                },
                "memory_used": {
                    "type": "integer"
                }
            }
        },
...
        "dto.UpdateClientRequest": {
            "description": "更新客户端配置的请求体",
            "type": "object",
            "properties": {
                "nickname": {
                    "type": "string",
                    "maxLength": 128,
                    "example": "My Client"
                },
                "rules": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.ProxyRule"
                    }
                }
            }
        },
...
        "handler.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "JWT Bearer token",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}